PhyloStreamerBackend/requirements.txt
aiofiles==23.1.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bio==1.4.0
biopython==1.79
biothings-client==0.2.6
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
msgpack==1.0.5
mygene==3.2.2
numpy==1.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygtrie==2.5.0
pyOpenSSL==23.1.1
pytz==2022.2.1
redis==4.5.4
requests==2.28.1
seqmagick==0.8.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.12
watchdog==3.0.0
zope.interface==6.0
